Multiplier mutation
- Reworked: Multiplier Mutation

+++ Catalano.Genetic/src/Catalano/Genetic/Mutation/MultiplierMutation.java
@@ -1,8 +1,25 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+// Catalano Genetic Library
+// The Catalano Framework
+//
+// Copyright © Diego Catalano, 2012-2016
+// diego.catalano at live.com
+//
+//
+//    This library is free software; you can redistribute it and/or
+//    modify it under the terms of the GNU Lesser General Public
+//    License as published by the Free Software Foundation; either
+//    version 2.1 of the License, or (at your option) any later version.
+//
+//    This library is distributed in the hope that it will be useful,
+//    but WITHOUT ANY WARRANTY; without even the implied warranty of
+//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+//    Lesser General Public License for more details.
+//
+//    You should have received a copy of the GNU Lesser General Public
+//    License along with this library; if not, write to the Free Software
+//    Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+//
+
 package Catalano.Genetic.Mutation;
 
 import Catalano.Genetic.Chromosome.DoubleChromosome;
@@ -10,26 +27,44 @@ import Catalano.Math.Tools;
 import java.util.Random;
 
 /**
- *
- * @author Diego
+ * Multiplier Mutation.
+ * 
+ * Support: Double chromosome.
+ * 
+ * @author Diego Catalano
  */
 public class MultiplierMutation implements IMutation<DoubleChromosome>{
     
-    private double balancer;
+    private float balancer;
 
+    /**
+     * Get balancer value.
+     * @return Balancer value.
+     */
     public double getBalancer() {
         return balancer;
     }
 
-    public void setBalancer(double balancer) {
+    /**
+     * Set balancer value.
+     * @param balancer Balancer value.
+     */
+    public void setBalancer(float balancer) {
         this.balancer = balancer;
     }
 
+    /**
+     * Initializes a new instance of the MultiplierMutation class.
+     */
     public MultiplierMutation() {
-        this(0.5);
+        this(0.5f);
     }
 
-    public MultiplierMutation(double balancer) {
+    /**
+     * Initializes a new instance of the MultiplierMutation class.
+     * @param balancer Balancer value.
+     */
+    public MultiplierMutation(float balancer) {
         this.balancer = balancer;
     }
